libibverbs: Add support for UET QPs
The UE transport is designed around connectionless communication.
To expose UET through verbs, we introduce a new reliable-
unconnected QP type (named to align with existing QP types).

As part of this change, the QP type is no longer automatically
coupled with a specific communication transport.  It is instead
treated as a communication semantic - e.g. RC, UD, UC, RU,
as originally intended.  Details on this separation is given
below.

When creating a QP, additional attributes may be needed. For
example, the provider may need to know what specific protocol
the QP will use.  The provider may also need full or partial
addressing information.  For instance, with UET, the QP may
be associated with a specific JobID.  There may be additional
information that a provider must know prior to allocating the
HW resources associated with the QP.

To support a more flexible QP creation model, we extend struct
ibv_qp_init_attr_ex to include struct ibv_qp_attr.  This
allows an application to provide more complete information
on the QP and its configuration during creation.  This has a
side effect that providers can use this information to
streamline QP creation and modification calls.

QP addressing data is defined as part of struct ibv_qp_attr.
That addressing is extended to support more generic (non-IB)
addressing formats.  Addressing information includes the
protocol stack to use with the QP.  This becomes necessary to
create a QP, with the underlying protocol identified separately.
For example, on a NIC which supports both RoCEv2 and UET,
a UD QP may run over either RoCEv2 or UET.  The extended QP
attributes allow the user to specify which protocol is in use,
as well as any necessary addressing data needed to setup
HW resources.

Signed-off-by: Sean Hefty <shefty@nvidia.com>