libibverbs: Allow using AHs to address UET peers
Support for job based addressing tables is provider specific
and optional.  As an alternative mechanism to address peers,
extend the functionality of using an address handle.

With legacy address handles, the AH identifies a remote port.
Information about the target QP is provided separately.

To support more generic addressing, we introduce a new call,
ibv_create_ah_ex().  This call accepts as input a job key
and the more generic struct ibv_addr_info.  This allows
the user to indicate what protocol the AH is targeting.

When used with a legacy QP / protocol, the behavior of
ibv_create_ah_ex() is similar to ibv_create_ah().  The
job key parameter must be NULL, and the returned AH still
refers to a destination port.

However, when used to create an AH for UET, the AH is
associated with the given job key (job ID), and it
references the target UET QP.  That is, the AH is job
restricted in its use.  Associating the AH with the
target QP allows HW implementation flexibility.  For
example, if the provider does support addressing tables,
but the application is using AHs, then an AH can map
to an entry in the table.  For providers which only
support the legacy AH concept, multiple SW AHs targeting
the same port may map to the same HW resource, with the
SW AHs storing the peer QPN.

A provider will be able to determine the structure of
the AH based on the QP that it is used with.

Signed-off-by: Sean Hefty <shefty@nvidia.com>