libibverbs: Report UET transport details in completions
Allow UET specific information to be reported as part of
work completions.  This includes the larger immediate
data size, the job ID carried in the transport header,
and a peer ID, also carried in the transport header.

Included with completion data is a UET transport field,
called the initiator in UEC terminology.  This is a
user configurable value intended to map to the rank
number for a parallel application.  The initiator field
only has meaning within a specific job ID.  As a result,
when the value is valid in a completion, so is the job ID.
(For UET, the initiator value is part of the UET address.)

The verbs naming of this field is the slightly more
generic term, peer ID.

Signed-off-by: Sean Hefty <shefty@nvidia.com>