libibverbs: Define UET capable memory registration
UET assigns additional properties and features that may
be associated with a MR.  To support a broader set of
features and allow greater flexibility in how a device
may implement memory registration, a new ibv_reg_mr_attr()
call is added.  This call allows for further extensibility.

A MR is no longer exclusively accessible by a single
protocol.  It's possible for a NIC to expose a MR that is
simultaneously usable with RDMA and UE transports, for
example.  A later patch will introduce new MR capabilities
to capture provider restrictions and features of MRs.

ibv_reg_mr_attr() takes an extensible input / output
struct ibv_mr_attr as a parameter.  The base behavior
of the call is similar to ibv_reg_mr().  However, the
call can also support registering memory backed by a
DMABuf (using an fd to identify the region).  If a
64-bit rkey is assigned to the retion, the rkey will
be returned through the attr parameter.  (For
compatibility reasons, struct ibv_mr is left unchanged.
As such, it can only report the lower 32-bits of the rkey).

A user may also request that any provider assigned rkey
be limited to 32-bits, allowing the rkey to be used
across both RDMA and UE transports.  A 32-bit rkey
may be converted to a 64-bit rkey by setting the upper
bits to 0.

A MR may be further restricted to only being accessible
by operations which carry a specific job ID.  To restrict
a MR to a job, a job key may be provided during registration.

Finally, ibv_reg_mr_attr() introduces a concept knows as
a derived MR.  Derived MRs are similar to legacy RDMA
memory windows.  Derived MRs are new MRs that are wholy
contained within an existnig MR (to share page mappings,
for example), but have different access rights or other
attributes (e.g. job ID assignment).  A derive_cnt field
indicates the number of derived MRs that an application
intends to create.  This field is considered an
optimization and may be ignored by the provider.

Unlike MWs, which use the send WR API to create a new
window, derived MRs use the memory registration API to
create a new MR.

Signed-off-by: Sean Hefty <shefty@nvidia.com>