rdma_cm/examples: Support multiple OOB connections
Add common code to support a server communicating with multiple
clients using OOB socket communication.  This will be used by
subsequent patches to allow scalability testing across multiple
processes.

To avoid confusion with follow on changes, we rename the server
as 'root' and refer to clients as 'leaf's.  Data transfers are
either up from leaf to root or down from the root to a leaf.

Signed-off-by: Sean Hefty <shefty@nvidia.com>