libibverbs: Add job id support
Job IDs are used to identify a distributed application.
The concept is widely used in HPC and AI applications, to
identify a set of distributed processes as belonging to
a single application.

Job IDs are integral to ultra ethernet.  A job ID is
carried in every transport message and is part of a
UET QP address.  UEC defines that job IDs must be managed
by a privileged entity.  The association of a job ID to
a specific QP is a protected operation.

A simple view of the job security model is shown as this
object model:

   device  <---  job ID
      ^           ^
      |           |
      PD   <---  job key
      ^ ^         ^
      |  \___     | (optional)
      QP      --- MR

This patch focuses on the job ID.  Job keys are discussed
in a following patch.

We define new verb calls to allocate a job object.  Each
job object is assigned a unique ID.  The assignment of
ID values to job objects it outside the scope of the API,
and would usually be handled through a job launcher or
process manager.  The ibv_alloc_job() call is use to
create and configure a job object.  It is expected that
the kernel will enforce that callers have the proper
privileges to create job objects on devices.  (Similar
to opening QP 0 or 1).

Once a job object has been created, it may be shared with
local processes using a shared fd mechanism.  The creating
process obtains a sharable fd using ibv_export_job() and
exchanges the fd with the processes of the job (e.g. via
sockets).  On receiving the fd, the processes use
ibv_import_job() to setup local job resources.

A job is associated with addressing information, which
includes protocol stack data, as well as an ID.  The number
of bytes of the ID which are valid is dependent on the
associated protocol.  For UET, it is 3-bytes.

A job object performs an additional function beyond linking
a QP with a job ID.  It defines a mechanism by which local
processes can share addressing information of peers.  This
can reduce the amount of memory used to store addresses
locally and enables future optimizations, such as applying
job level encryption.  The feature will also map well to
HPC and AI applications that identify peers using a rank.

Conceptually, a virtual address array may be stored with
a job object.  Addresses are inserted or removed from the
array at a given index location.  The intent is that the
index can map directly to the process' rank.  When sending
to a peer, the peer can be identified by the job plus the
index.

Note that the implementation for the job's addressing array
is not defined.  A vendor may implement this in a variety
of ways.  Addresses may be pre-inserted by the job launcher,
and the transport addresses may be generated using an
algorithm.

Signed-off-by: Sean Hefty <shefty@nvidia.com>