libibverbs: Allow provider to describe immediate data limits
Legacy RDMA transports are restricted to 32-bits of immediate
data, while UET supports 64-bits.  Additionally, UET does not
require that RDMA writes with immediate consume a posted
receive buffer at the target.  The spec even goes so far as
to mandate that RDMA traffic be treated separately at the
target than send operations, but such a mandate is not
visible in the transport and places restrictions on the NIC
implementat.

NICs that support multiple protocols may be optimized for
legacy RDMA support.  We add immediate data attributes to the
protocol capabilities.  A provider can indicate the
maximum size, in bytes, of imm data supported by the device
for the specified protocol.

Truncated immediate data is extended with 0s when writing
to the wire, and completions report the lowest valid bits.

An imm data capability flag is also defined, IMM_DATA_CAP_RQ.
When this flag is set, RDMA writes with immediate will
consume a posted receive buffer on the QP.  This is the
standard behavior for legacy RDMA.

Signed-off-by: Sean Hefty <shefty@nvidia.com>