libibverbs: Define memory registration capabilities
Add memory registration capabilities to the protocol
caps.  MR caps include the following:

The maximum size of an rkey, in bytes, supported by
the provider. This value may be smaller than that
carried by the transport.  In such cases, the upper
bits of the rkey will be 0.

IBV_MR_QP_ATTRACH: This flag indicates that MRs are
not accessible until they are explicitly attached to a
QP.  Once registered, a MR may be attached or detached
from a QP by posting a new IBV_WR_ATTACH_MR or
IBV_WR_DETACH_MR operation on the QP.  MRs referenced
by the lkey(s) in the correponding sg_list will be
accessible once attached, or no longer accessible once
detached.  If this flag is set to 0, MRs are immediately
accessible by any QP sharing the PD as the MR.
    
IBV_MR_USER_RKEY: This flag indicates that the provider
supports the ability of the caller to specify the rkey
to assign to the MR.

Signed-off-by: Sean Hefty <shefty@nvidia.com>